cleanup introspect
